requirements.txt
asgiref==3.8.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==5.1.6
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==11.1.0
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.6.3
